refactor(tron): centralize approval sponsoring extension
