+++ typescript/packages/mechanisms/tron/src/exact/client/index.ts
@@ -6,4 +6,3 @@ export {
   getPermit2AllowanceReadParams,
   type Permit2AllowanceParams,
 } from "./permit2Helpers";
-export { trySignTrc20ApprovalResourceSponsoringExtension } from "./trc20approval";
