+++ typescript/packages/mechanisms/tron/src/exact/facilitator/index.ts
@@ -1,38 +1,3 @@
 export { ExactTronScheme } from "./scheme";
 export { registerExactTronScheme } from "./register";
 export type { TronFacilitatorConfig } from "./register";
-export {
-  buildTrc20ApprovalSponsoringRequest,
-  decodeSignedTrc20Approval,
-  validateTrc20ApprovalForPayment,
-  type DecodedTrc20Approval,
-  type Trc20ApprovalValidationOptions,
-  type Trc20ApprovalValidationResult,
-} from "./trc20approval";
-export {
-  buildTrc20SponsoringPlan,
-  createTrc20ApprovalResourceSponsoringRuntime,
-  createTrc20ResourceSponsoringRuntime,
-  createStaticTrc20ResourceSponsoringPolicy,
-  createTronWebResourceSponsoringChain,
-  InMemoryTrc20SponsoringCoordinator,
-  resourceUnitsToStakeSun,
-  type InMemoryTrc20SponsoringCoordinatorOptions,
-  type CreateTrc20ResourceSponsoringRuntimeOptions,
-  type StaticTrc20ResourceSponsoringPolicyOptions,
-  type ManagedTrc20ApprovalResourceSponsoringRuntime,
-  type PreparedTronAction,
-  type Trc20ResourceLeg,
-  type Trc20ResourceSponsoringChain,
-  type Trc20ResourceSponsoringPolicy,
-  type Trc20ResourceSponsoringRuntimeOptions,
-  type Trc20SponsoringCoordinator,
-  type Trc20SponsoringOperation,
-  type Trc20SponsoringPlan,
-  type Trc20SponsoringPreflight,
-  type TronResourceSnapshot,
-  type TronResourceOwnerActionIntent,
-  type TronResourceOwnerSigner,
-  type TronResourceType,
-  type TronWebResourceSponsoringChainOptions,
-} from "../../resource-sponsoring";
