+++ typescript/packages/mechanisms/tron/src/exact/index.ts
@@ -1,12 +1,1 @@
 export { ExactTronScheme } from "./client/scheme";
-export {
-  TRC20_APPROVAL_MAX_AMOUNT,
-  TRC20_APPROVAL_RESOURCE_SPONSORING_KEY,
-  TRC20_APPROVAL_RESOURCE_SPONSORING_VERSION,
-  type Trc20ApprovalResourceSponsoringFacilitatorExtension,
-  type Trc20ApprovalResourceSponsoringInfo,
-  type Trc20ApprovalResourceSponsoringRequest,
-  type Trc20ApprovalResourceSponsoringRuntime,
-  type Trc20SponsorshipExecutionOptions,
-  type Trc20SponsorshipRevalidationResult,
-} from "./extensions";
