+++ typescript/packages/mechanisms/tron/src/shared/extensions/index.ts
@@ -4,3 +4,12 @@ export {
   type Trc20ApprovalExtensionAttempt,
 } from "./resourceSponsoring";
 export { executeTrc20Sponsorship, verifyTrc20Sponsorship } from "./trc20ApprovalResourceSponsoring";
+export * from "./trc20ApprovalContract";
+export {
+  buildTrc20ApprovalSponsoringRequest,
+  decodeSignedTrc20Approval,
+  validateTrc20ApprovalForPayment,
+  type DecodedTrc20Approval,
+  type Trc20ApprovalValidationOptions,
+  type Trc20ApprovalValidationResult,
+} from "./trc20ApprovalTransaction";
