+++ typescript/packages/mechanisms/tron/test/unit/trc20-resource-sponsoring-runtime.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it, vi } from "vitest";
-import type { Trc20ApprovalResourceSponsoringRequest } from "../../src/exact/extensions";
+import type { Trc20ApprovalResourceSponsoringRequest } from "../../src/shared/extensions/trc20ApprovalContract";
 import {
   buildTrc20SponsoringPlan,
   createTrc20ApprovalResourceSponsoringRuntime,
