+++ typescript/packages/mechanisms/tron/test/unit/trc20-resource-sponsoring-tronweb.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 import { TronWeb, utils as tronUtils } from "tronweb";
-import type { Trc20ApprovalResourceSponsoringRequest } from "../../src/exact/extensions";
+import type { Trc20ApprovalResourceSponsoringRequest } from "../../src/shared/extensions/trc20ApprovalContract";
 import {
   createTronWebResourceSponsoringChain,
   type TronResourceOwnerSigner,
